+++ packages/api/tests/lambdas/ems-distribution-report.js
@@ -6,17 +6,17 @@ const path = require('path');
 const test = require('ava');
 const { aws } = require('@cumulus/common');
 const { testUtils: { randomString } } = require('@cumulus/common');
-const { generateAndStoreDistributionReport } = require('../../lambdas/ems-distribution-report');
+const { bucketsPrefixes, generateAndStoreDistributionReport } = require('../../lambdas/ems-distribution-report');
 const models = require('../../models');
 const { fakeGranuleFactory, fakeFileFactory } = require('../../lib/testUtils');
-
-process.env.system_bucket = 'test-bucket';
-process.env.stackName = 'test-stack';
-process.env.ems_provider = 'testEmsProvider';
 
 let granuleId;
 
 test.before(async () => {
+  process.env.system_bucket = randomString();
+  process.env.stackName = 'test-stack';
+  process.env.ems_provider = 'testEmsProvider';
+
   process.env.GranulesTable = randomString();
   process.env.FilesTable = randomString();
 
@@ -44,8 +44,9 @@ test.after.always(async () => {
 
 
 test.beforeEach(async (t) => {
-  // Create the internal bucket
-  t.context.internalBucket = randomString();
+  t.context.internalBucket = process.env.system_bucket;
+  const { logsBucket, logsPrefix } = bucketsPrefixes();
+
   await aws.s3().createBucket({ Bucket: t.context.internalBucket }).promise();
 
   // Read in all of the server logs from the fixtures files
@@ -55,11 +56,10 @@ test.beforeEach(async (t) => {
     fs.readFile(path.join(fixturesDirectory, serverFilename), 'utf8')));
 
   // Upload the S3 server logs to the internal bucket
-  t.context.logsPrefix = randomString();
   await Promise.all(serverLogs.map((serverLog) =>
     aws.s3().putObject({
-      Bucket: t.context.internalBucket,
-      Key: aws.s3Join([t.context.logsPrefix, `${randomString()}.log`]),
+      Bucket: logsBucket,
+      Key: aws.s3Join([logsPrefix, `${randomString()}.log`]),
       Body: serverLog
     }).promise()));
 });
@@ -69,14 +69,7 @@ test.afterEach.always(async (t) => {
 });
 
 test.serial('emsDistributionReport writes a correct report out to S3 when no previous reports exist', async (t) => {
-  const logsBucket = t.context.internalBucket;
-  const logsPrefix = t.context.logsPrefix;
-
   const reportsBucket = t.context.internalBucket;
-  const reportsPrefix = randomString();
-
-  const provider = randomString();
-  const stackName = randomString();
 
   const reportStartTime = moment.utc('1981-06-01T01:00:00Z');
   const reportEndTime = moment.utc('1981-06-01T15:00:00Z');
@@ -84,13 +77,7 @@ test.serial('emsDistributionReport writes a correct report out to S3 when no pre
   // Generate the distribution report
   const report = await generateAndStoreDistributionReport({
     reportStartTime,
-    reportEndTime,
-    logsBucket,
-    logsPrefix,
-    reportsBucket,
-    reportsPrefix,
-    stackName,
-    provider
+    reportEndTime
   });
 
   // Fetch the distribution report from S3
@@ -112,20 +99,12 @@ test.serial('emsDistributionReport writes a correct report out to S3 when no pre
 });
 
 test.serial('emsDistributionReport writes a correct report out to S3 when one report already exists', async (t) => {
-  const logsBucket = t.context.internalBucket;
-  const logsPrefix = t.context.logsPrefix;
-
-  const reportsBucket = t.context.internalBucket;
-  const reportsPrefix = randomString();
-
-  const provider = randomString();
-  const stackName = randomString();
-
   const reportStartTime = moment.utc('1981-06-01T01:00:00Z');
   const reportEndTime = moment.utc('1981-06-01T15:00:00Z');
 
   const reportName = `${reportStartTime.format('YYYYMMDD')}_${process.env.ems_provider}_DistCustom_${process.env.stackName}.flt`;
 
+  const { reportsBucket, reportsPrefix } = bucketsPrefixes();
   await aws.s3().putObject({
     Bucket: reportsBucket,
     Key: aws.s3Join([reportsPrefix, reportName]),
@@ -135,13 +114,7 @@ test.serial('emsDistributionReport writes a correct report out to S3 when one re
   // Generate the distribution report
   await generateAndStoreDistributionReport({
     reportStartTime,
-    reportEndTime,
-    logsBucket,
-    logsPrefix,
-    reportsBucket,
-    reportsPrefix,
-    stackName,
-    provider
+    reportEndTime
   });
 
   // Fetch the distribution report from S3
@@ -163,17 +136,10 @@ test.serial('emsDistributionReport writes a correct report out to S3 when one re
 });
 
 test.serial('emsDistributionReport writes a correct report out to S3 when two reports already exist', async (t) => {
-  const logsBucket = t.context.internalBucket;
-  const logsPrefix = t.context.logsPrefix;
-
-  const reportsBucket = t.context.internalBucket;
-  const reportsPrefix = randomString();
-
-  const provider = randomString();
-  const stackName = randomString();
-
   const reportStartTime = moment.utc('1981-06-01T01:00:00Z');
   const reportEndTime = moment.utc('1981-06-01T15:00:00Z');
+
+  const { reportsBucket, reportsPrefix } = bucketsPrefixes();
 
   const reportName = `${reportStartTime.format('YYYYMMDD')}_${process.env.ems_provider}_DistCustom_${process.env.stackName}.flt`;
 
@@ -193,13 +159,7 @@ test.serial('emsDistributionReport writes a correct report out to S3 when two re
   // Generate the distribution report
   await generateAndStoreDistributionReport({
     reportStartTime,
-    reportEndTime,
-    logsBucket,
-    logsPrefix,
-    reportsBucket,
-    reportsPrefix,
-    stackName,
-    provider
+    reportEndTime
   });
 
   // Fetch the distribution report from S3
